requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9
django-redis==5.4.0
python-dotenv==1.0.1
gunicorn==22.0.0
setuptools==75.0.0
whitenoise==6.7.0